--- keyboards/cheapino/matrix-encoder.h
@@ -1,6 +1,0 @@
-//
-// Created by Thomas Haukland on 25/03/2023.
-//
-#include "quantum.h"
-
-void fix_encoder_action(matrix_row_t current_matrix[]);
