--- keyboards/cheapino/matrix-ghosting.h
@@ -1,5 +1,0 @@
-//
-// Created by Thomas Haukland on 2024-05-05.
-//
-
-void fix_ghosting(matrix_row_t current_matrix[]);
