+++ keyboards/cheapino/rules.mk
@@ -1,3 +1,1 @@
-SRC += matrix-encoder.c
-SRC += matrix-ghosting.c
 SRC += matrix.c
